Initial Code Upload into Github.
Value Add:
It will allow me to better save my code into my Github for future operations and usage.

Use Cases:
This upload enables future commits and development of the program.
Enables to Programmer and Public individuals to observe and see the initial upload of the code.

Rationale for Choices:
The code and commit seen here are the primary option to be able to get the code officially uploaded and readied. I have considered that there are possible other options to saving the code, such as just to my main computer, however this has several drawbacks. Primarily locking my ability to access the code from other machines and thus limiting my ability to improve the code further and not allowing others to view or access the code.