Adjusted logarithm calculator to now incorporate the functionality to calculate Base-10 Logarithm or do Natural Logarithm.
Value Add:
Users now can better utilise the features of the program to now calculate the natural log OR base-10 log of a specific number and choose which one to use. 

Use Cases:
Users can better use this better calculate the natural log of a number and be able to calculate the base-10 for the same number if ran again.

Rationale for Choices:
When adding onto the code, I found it more viable and simplier to combine the function area for the logarithm calculations to incorporate both the Natural Log and Base-10 Log, to both compact the design for easier compiling, as well as simplifying the process.