Made a heavy vision of the calculator to change the overall functionality to where now you can either use power or log in order to reduce complexity and improve runtime.
Value Add:
Users can now have a much easier time running the program to either just use power functions or log functions to find their answer for various numbers and inputs. 

Use Cases:
Teachers using this program can now have a easier time running the program to calculate either the power of a number or find the log of a number and whether to use Natural Log or Base-10 Log.

Rationale for Choices:
When adjusting and rewriting the code, I choose to do this in order to improve the running capability of the program and no longer force people to calculate a power number and be forced to calculate a log number. This was done to give users far more control over the program to find their necessary solutions and quickly be able to get their answer without sitting through a long process.